+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+7.3
+---
+ * Add the `BackedEnumValue` constraint
+
 7.2
 ---